+++ ninja-servlet-jpa-blog-integration-test/src/main/java/conf/Routes.java
@@ -79,7 +79,8 @@ public class Routes implements ApplicationRoutes {
         ///////////////////////////////////////////////////////////////////////
         // Assets (pictures / javascript)
         ///////////////////////////////////////////////////////////////////////    
-        router.GET().route("/assets/.*").with(AssetsController.class, "serve");
+        router.GET().route("/assets/webjars/{fileName: .*}").with(AssetsController.class, "serveWebjars");
+        router.GET().route("/assets/{fileName: .*}").with(AssetsController.class, "serveStatic");
         
         ///////////////////////////////////////////////////////////////////////
         // Index / Catchall shows index page